move calculate_score_groups
