+++ alphadia/plexscoring/utils.py
@@ -264,3 +264,141 @@ def merge_missing_columns(
 
     # merge
     return left_df.merge(right_df[on + missing_from_left], on=on, how=how)
+
+
+def calculate_score_groups(
+    input_df: pd.DataFrame,
+    group_channels: bool = False,
+):
+    """
+    Calculate score groups for DIA multiplexing.
+
+    On the candidate selection level, score groups are used to group ions across channels.
+    On the scoring level, score groups are used to group channels of the same precursor and rank together.
+
+    This function makes sure that all precursors within a score group have the same `elution_group_idx`, `decoy` status and `rank` if available.
+    If `group_channels` is True, different channels of the same precursor will be grouped together.
+
+    Parameters
+    ----------
+
+    input_df : pandas.DataFrame
+        Precursor dataframe. Must contain columns 'elution_group_idx' and 'decoy'. Can contain 'rank' column.
+
+    group_channels : bool
+        If True, precursors from the same elution group will be grouped together while seperating different ranks and decoy status.
+
+    Returns
+    -------
+
+    score_groups : pandas.DataFrame
+        Updated precursor dataframe with score_group_idx column.
+
+    Example
+    -------
+
+    A precursor with the same `elution_group_idx` might be grouped with other precursors if only the `channel` is different.
+    Different `rank` and `decoy` status will always lead to different score groups.
+
+    .. list-table::
+        :widths: 25 25 25 25 25 25
+        :header-rows: 1
+
+        * - elution_group_idx
+          - rank
+          - decoy
+          - channel
+          - group_channels = False
+          - group_channels = True
+
+        * - 0
+          - 0
+          - 0
+          - 0
+          - 0
+          - 0
+
+        * - 0
+          - 0
+          - 0
+          - 4
+          - 1
+          - 0
+
+        * - 0
+          - 1
+          - 0
+          - 0
+          - 2
+          - 1
+
+        * - 0
+          - 1
+          - 1
+          - 0
+          - 3
+          - 2
+
+    """
+
+    @nb.njit(cache=USE_NUMBA_CACHING)
+    def channel_score_groups(elution_group_idx, decoy, rank):
+        """
+        Calculate score groups for channel grouping.
+
+        Parameters
+        ----------
+
+        elution_group_idx : numpy.ndarray
+            Elution group indices.
+
+        decoy : numpy.ndarray
+            Decoy status.
+
+        rank : numpy.ndarray
+            Rank of precursor.
+
+        Returns
+        -------
+
+        score_groups : numpy.ndarray
+            Score groups.
+        """
+        score_groups = np.zeros(len(elution_group_idx), dtype=np.uint32)
+        current_group = 0
+        current_eg = elution_group_idx[0]
+        current_decoy = decoy[0]
+        current_rank = rank[0]
+
+        for i in range(len(elution_group_idx)):
+            # if elution group, decoy status or rank changes, increase score group
+            if (
+                (elution_group_idx[i] != current_eg)
+                or (decoy[i] != current_decoy)
+                or (rank[i] != current_rank)
+            ):
+                current_group += 1
+                current_eg = elution_group_idx[i]
+                current_decoy = decoy[i]
+                current_rank = rank[i]
+
+            score_groups[i] = current_group
+        return score_groups
+
+    # sort by elution group, decoy and rank
+    # if no rank is present, pretend rank 0
+    if "rank" in input_df.columns:
+        input_df = input_df.sort_values(by=["elution_group_idx", "decoy", "rank"])
+        rank_values = input_df["rank"].values
+    else:
+        input_df = input_df.sort_values(by=["elution_group_idx", "decoy"])
+        rank_values = np.zeros(len(input_df), dtype=np.uint32)
+
+    if group_channels:
+        input_df["score_group_idx"] = channel_score_groups(
+            input_df["elution_group_idx"].values, input_df["decoy"].values, rank_values
+        )
+    else:
+        input_df["score_group_idx"] = np.arange(len(input_df), dtype=np.uint32)
+
+    return input_df.sort_values(by=["score_group_idx"]).reset_index(drop=True)
